[aiEnhance] refactor
